Master

See merge request compsci308_2020spring/parser_parser_team02!31

+++ src/slogo/model/Commands/TurtleCommands/TurtleCommand.java
@@ -3,10 +3,10 @@ package slogo.model.Commands.TurtleCommands;
 import slogo.model.Commands.Command;
 
 public abstract class TurtleCommand extends Command {
-  private static final int FACING_RIGHT = 90;
-  private static final int FACING_UP = 360;
-  private static final int FACING_LEFT = 270;
-  private static final int FACING_DOWN = 180;
+  private static final double FACING_RIGHT = 90;
+  private static final double FACING_UP = 360;
+  private static final double FACING_LEFT = 270;
+  private static final double FACING_DOWN = 180;
   protected int forward = 1;
   protected int backward = -1;
 
@@ -14,32 +14,32 @@ public abstract class TurtleCommand extends Command {
     super(name);
   }
 
-  protected double degreesToRadians(int angle) {
+  protected double degreesToRadians(double angle) {
     return angle * Math.PI / 180;
   }
 
-  protected void moveTurtleTo(int x, int y) {
+  protected void moveTurtleTo(double x, double y) {
     turtle.setX(x);
     turtle.setY(y);
   }
 
-  protected boolean facingTopRight(int angle) {
+  protected boolean facingTopRight(double angle) {
     return angle >= 0 && angle < FACING_RIGHT;
   }
 
-  protected boolean facingTopLeft(int angle) {
+  protected boolean facingTopLeft(double angle) {
     return angle >= FACING_LEFT && angle < FACING_UP;
   }
 
-  protected boolean facingBottomRight(int angle) {
+  protected boolean facingBottomRight(double angle) {
     return angle >= FACING_RIGHT && angle < FACING_DOWN;
   }
 
-  protected boolean facingBottomLeft(int angle) {
+  protected boolean facingBottomLeft(double angle) {
     return angle >= FACING_DOWN && angle < FACING_LEFT;
   }
 
-  protected int getAdjustedAngle(int angle) {
+  protected double getAdjustedAngle(double angle) {
     if (facingTopRight(angle))
       return angle;
     else if (facingTopLeft(angle))
@@ -50,39 +50,39 @@ public abstract class TurtleCommand extends Command {
       return angle - FACING_DOWN;
   }
 
-  protected int getXMultiplier(int angle) {
+  protected int getXMultiplier(double angle) {
     if (angle >= 0 && angle < FACING_DOWN)
       return 1;
     else
       return -1;
   }
 
-  protected int getYMultiplier(int angle) {
+  protected int getYMultiplier(double angle) {
     if (angle >= 90 && angle < 270)
       return -1;
     else
       return 1;
   }
 
-  protected void moveTurtle(int directionMultiplier, int distance) {
-    int angle = getAdjustedAngle(turtle.getDegree());
+  protected void moveTurtle(int directionMultiplier, double distance) {
+    double angle = getAdjustedAngle(turtle.getDegree());
     int xMultiplier = directionMultiplier * getXMultiplier(angle);
     int yMultiplier = directionMultiplier * getYMultiplier(angle);
 
     double angleToRadians = degreesToRadians(angle);
     double rightAngle = degreesToRadians(FACING_RIGHT);
 
-    int newX = turtle.getX() + xMultiplier * (int) (distance * Math.sin(angleToRadians));
-    int newY = turtle.getY() + yMultiplier * (int) (distance * Math.sin(rightAngle - angleToRadians));
+    double newX = turtle.getX() + xMultiplier * (distance * Math.sin(angleToRadians));
+    double newY = turtle.getY() + yMultiplier * (distance * Math.sin(rightAngle - angleToRadians));
 
     turtle.setX(newX);
     turtle.setY(newY);
   }
 
-  protected void rotateTurtle(int direction, int degrees) {
-    int currentDegrees = turtle.getDegree();
+  protected void rotateTurtle(int direction, double degrees) {
+    double currentDegrees = turtle.getDegree();
 
-    int newDegrees = 0;
+    double newDegrees = 0;
     if (direction == 1)
       newDegrees = (currentDegrees + degrees) % 360;
     else {
@@ -100,7 +100,7 @@ public abstract class TurtleCommand extends Command {
     turtle.setDegree(newDegrees);
   }
 
-  protected double distanceFormula(int x1, int y1, int x2, int y2) {
+  protected double distanceFormula(double x1, double y1, double x2, double y2) {
     return Math.sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
   }
